move: 세션 관련 객체 패키지 이동
